change structure again

+++ scripts/reconstruction.py
@@ -32,6 +32,7 @@ no_debugg = False
 # Updating config file
 up_from_ini(file_config, inifile)
 
+
 # Checking that the confing file has the right number of patternRecognition and GBLAlign iterations
 with open(file_config) as cfgfile:
     data = cfgfile.read()
@@ -40,11 +41,13 @@ with open(file_config) as cfgfile:
 cfgfile.close()
 
 for run_number in run_list:
+
     # Creating runlist with gear files to be used
     gears = name_georun(basic_geo, run_number, iterations)
 
     # Creating directory to store gear files per run
     gears_dir = geometry_files_path + "run_%s" % run_number
+    print(gears)
     try:
         check_call("mkdir " + gears_dir, shell=True)
     except CalledProcessError:
@@ -60,29 +63,33 @@ for run_number in run_list:
     # +-----------------------------------------------------------
 
     # Running noisypixel, clustering and hitmaker
-    if True:
+    if False:
         runlist_builder(runlist_path + runlist_file, gears[0], run_number, beam_energy, thr)
         run_recon(config_path + config_file, runlist_path + runlist_file, "noisypixel", run_number)
         run_recon(config_path + config_file, runlist_path + runlist_file, "clustering", run_number)
         run_recon(config_path + config_file, runlist_path + runlist_file, "hitmaker", run_number)
 
     # Running pattern recognition and GBLAlign iteration 1
-    if True:
+    if False:
         runlist_builder(runlist_path + runlist_file, gears[1], run_number, beam_energy, thr)
         run_recon(config_path + config_file, runlist_path + runlist_file, "patternRecognition1", run_number)
         run_recon(config_path + config_file, runlist_path + runlist_file, "GBLAlign1", run_number)
+        runlist_builder(runlist_path + runlist_file, gears[2], run_number, beam_energy, thr)
+        run_recon(config_path + config_file, runlist_path + runlist_file, "patternRecognition1", run_number)
 
     # Running pattern recognition and GBLAlign iteration 2
     if False:
-        runlist_builder(runlist_path + runlist_file, gears[2], run_number, beam_energy, thr)
         run_recon(config_path + config_file, runlist_path + runlist_file, "patternRecognition2", run_number)
         run_recon(config_path + config_file, runlist_path + runlist_file, "GBLAlign2", run_number)
+        runlist_builder(runlist_path + runlist_file, gears[3], run_number, beam_energy, thr)
+        run_recon(config_path + config_file, runlist_path + runlist_file, "patternRecognition2", run_number)
 
     # Running pattern recognition and GBLAlign iteration 3
     if False:
-        runlist_builder(runlist_path + runlist_file, gears[3], run_number, beam_energy, thr)
         run_recon(config_path + config_file, runlist_path + runlist_file, "patternRecognition3", run_number)
         run_recon(config_path + config_file, runlist_path + runlist_file, "GBLAlign3", run_number)
+        runlist_builder(runlist_path + runlist_file, gears[4], run_number, beam_energy, thr)
+        run_recon(config_path + config_file, runlist_path + runlist_file, "patternRecognition3", run_number)
 
     # Running pattern recognition and GBLAlign iteration 4
     if False:
@@ -109,9 +116,11 @@ for run_number in run_list:
                     print("Good alignment!! continuing")
 
     # Cleaning and organizing geometry directory
-    try:
-        check_call("rm " + geometry_files_path + gears[0], shell=True)
-        check_call("mv " + geometry_files_path + gears[1] + " " + geometry_files_path + "run_%s" % run_number + "/", shell=True)
-        check_call("mv " + geometry_files_path + gears[2] + " " + geometry_files_path + "run_%s" % run_number + "/", shell=True)
-    except CalledProcessError:
-        print("")
+    organize = True
+    if organize:
+        try:
+            check_call("rm " + geometry_files_path + gears[0], shell=True)
+            for file in range(1, len(gears)+1):
+                check_call("mv " + geometry_files_path + gears[file] + " " + geometry_files_path + "run_%s" % run_number + "/", shell=True)
+        except CalledProcessError:
+            print("")
